requirements.txt
selenium==3.141.0
peewee==3.14.1
prefect
webdriver-manager
python-dotenv